+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Changed
+- Deterministic function name ordering
+
+## Added
+- `lazy_deployment_completed` and `get_abi_ipfs` to known methods
+
 ## [0.0.11] - 2023-03-27
 
 ## Fixed